Fixed error in parsing PIN

+++ src/main/java/kmalfa/utils/PinCodeAnalyzer.java
@@ -1,22 +1,22 @@
 package kmalfa.utils;
 
 public class PinCodeAnalyzer {
-    private static final int[] CONSTANTS = {231, 455, 892, 351, 952, 116};
+    private static final float[] CONSTANTS = {231.f, 455.f, 892.f, 351.f, 952.f, 116.f};
 
     public static int getPin (int pin, int operation) {
         int newPin = pin;
         switch (operation) {
             case 1:
-                newPin = CONSTANTS[0] * (CONSTANTS[2] - pin);
+                newPin = (int)(CONSTANTS[0] * (CONSTANTS[2] - pin));
                 break;
             case 2:
-                newPin = CONSTANTS[3] / (CONSTANTS[5] + pin);
+                newPin = (int)(CONSTANTS[3] / (CONSTANTS[5] + pin));
                 break;
             case 3:
-                newPin = CONSTANTS[4] + CONSTANTS[1] - pin;
+                newPin = (int)(CONSTANTS[4] + CONSTANTS[1] - pin);
                 break;
             case 4:
-                newPin = CONSTANTS[2] - CONSTANTS[5] + pin;
+                newPin = (int)(CONSTANTS[2] - CONSTANTS[5] + pin);
                 break;
             //In future may be more operations
         }
@@ -28,16 +28,16 @@ public class PinCodeAnalyzer {
         int codedPin = actualPin;
         switch (operation) {
             case 1:
-                codedPin = CONSTANTS[2] - (actualPin / CONSTANTS[0]);
+                codedPin = (int)(CONSTANTS[2] - (actualPin / CONSTANTS[0]));
                 break;
             case 2:
-                codedPin = (CONSTANTS[3] / actualPin) - CONSTANTS[5];
+                codedPin = (int)((CONSTANTS[3] / actualPin) - CONSTANTS[5]);
                 break;
             case 3:
-                codedPin = CONSTANTS[4] + CONSTANTS[1] - actualPin;
+                codedPin = (int)(CONSTANTS[4] + CONSTANTS[1] - actualPin);
                 break;
             case 4:
-                codedPin = actualPin - CONSTANTS[2] + CONSTANTS[5];
+                codedPin = (int)(actualPin - CONSTANTS[2] + CONSTANTS[5]);
                 break;
             //In future may be more operations
         }
